requirements.txt
torch
torchvision
Pillow
numpy >= 1.20.0, < 2.0.0
pytorch-fid
fastapi[standard]
streamlit